extract interface for buffersegment and bufferqueue
