+++ phoenix-core/src/main/java/org/apache/phoenix/iterate/OrderedResultIterator.java
@@ -207,7 +207,7 @@ public class OrderedResultIterator implements PeekingResultIterator {
         List<Expression> expressions = Lists.newArrayList(Collections2.transform(orderByExpressions, TO_EXPRESSION));
         final Comparator<ResultEntry> comparator = buildComparator(orderByExpressions);
         try{
-            final DeferredByteBufferSortedQueue queueEntries = new DeferredByteBufferSortedQueue(comparator, limit,
+            final MappedByteBufferSortedQueue queueEntries = new MappedByteBufferSortedQueue(comparator, limit,
                     thresholdBytes);
             resultIterator = new PeekingResultIterator() {
                 int count = 0;
